refactor: moved each queries on different files
