+++ project/lib/db/db.ts
@@ -1,0 +1,5 @@
+import { config } from "dotenv";
+import { drizzle } from "drizzle-orm/neon-serverless";
+
+config({ path: ".env" });
+export const db = drizzle(process.env.DATABASE_URL!);
